Don't mangle "initSqlJs" function during minification/optimization
The code in sqljs has some custom logic on how the function should be
exported, which I guess breaks the minifier causing it to emit invalid
code when trying to invoke the function from the extension code.
The easiest workaround I found is to exclude the function name from
mangling, which preserves the original name and, when invoked, it
still works.

Fixes #36